no internetconnection before start.sh starts

+++ install/scripts/startapp.sh
@@ -1,18 +1,7 @@
-#!/bin/bash
+#!/bin/bash -x
 
 DOCKER_IMAGE='rhoerbe/pvzd-client-app'
-
-
-wget -q --tries=10 --timeout=20 --spider http://www.identinetics.com/
-if [[ $? -eq 0 ]]; then
-        echo "Online"
-        notify-send "Online - Preparing download"
-else
-        echo "Offline"
-        notify-send "Offline - Please connect to internet and start this script again"
-        break
-fi
-
+sleep 30
 
 runopt='-it'
 while getopts ":hpt" opt; do
@@ -38,40 +27,55 @@ if [ $(id -u) -ne 0 ]; then
     sudo="sudo"
 fi
 
-notify-send "Pulling docker image $DOCKER_IMAGE; please wait, Update can be bigger" -t 50000
-logger -p local0.info "pulling docker image $DOCKER_IMAGE"
-$sudo docker pull $DOCKER_IMAGE
-notify-send "Docker image $DOCKER_IMAGE up-to date; starting container"
 
-logger -p local0.info "mapping container user's home to $DOCKERDATA_DIR"
-$sudo mkdir -p $DATA_DIR/home/liveuser/
-$sudo chown -R liveuser:liveuser $DATA_DIR/home/liveuser/
+wget -q --tries=10 --timeout=20 --spider http://www.identinetics.com/
+if [[ $? -eq 0 ]]; then
+        echo "Online"
+        notify-send "Online - Preparing download"
+        logger -p local0.info "Online prepareing download"
 
-CONTAINERNAME='x11-app'
+        notify-send "Pulling docker image $DOCKER_IMAGE; please wait, Update has several 100MB " -t 50000
+        logger -p local0.info "pulling docker image $DOCKER_IMAGE"
+        $sudo docker pull $DOCKER_IMAGE
+        notify-send "Docker image $DOCKER_IMAGE up-to date; starting container"
 
-# remove dangling container
-if $sudo docker ps -a | grep $CONTAINERNAME > /dev/null; then
-    logger -p local0.info "deleting dangling container $CONTAINERNAME"
-    $sudo docker rm $CONTAINERNAME
+        logger -p local0.info "mapping container user's home to $DOCKERDATA_DIR"
+        $sudo mkdir -p $DATA_DIR/home/liveuser/
+        $sudo chown -R liveuser:liveuser $DATA_DIR/home/liveuser/
+
+        CONTAINERNAME='x11-app'
+
+        # remove dangling container
+        if $sudo docker ps -a | grep $CONTAINERNAME > /dev/null; then
+            logger -p local0.info "deleting dangling container $CONTAINERNAME"
+            $sudo docker rm $CONTAINERNAME
+        fi
+
+        ENVSETTINGS="
+            -e DISPLAY=$DISPLAY
+            -e http_proxy=$http_proxy
+            -e https_proxy=$https_proxy
+            -e HTTP_PROXY=$HTTP_PROXY
+            -e HTTPS_PROXY=$HTTPS_PROXY
+            -e no_proxy=$no_proxy
+        "
+        LOGSETTINGS="--log-driver=syslog --log-opt syslog-facility=local0"
+        VOLMAPPING="
+            --privileged -v /dev/bus/usb:/dev/bus/usb
+            -v /tmp/.X11-unix/:/tmp/.X11-unix:Z
+            -v $DATA_DIR/home/liveuser/:/home/liveuser:Z
+        "
+
+        logger -p local0.info "starting docker image $DOCKER_IMAGE"
+        notify-send "starting docker image $DOCKER_IMAGE"
+        $sudo docker run $runopt --rm \
+            --hostname=$CONTAINERNAME --name=$CONTAINERNAME \
+            $ENVSETTINGS $LOGSETTINGS $VOLMAPPING \
+            $DOCKER_IMAGE
+
+else
+        echo "Not pulling docker image - OFFLINE"
+        notify-send "Not pulling docker image - OFFLINE"
+        logger -p local0.info "Not pulling docker image - OFFLINE"
+
 fi
-
-ENVSETTINGS="
-    -e DISPLAY=$DISPLAY
-    -e http_proxy=$http_proxy
-    -e https_proxy=$https_proxy
-    -e HTTP_PROXY=$HTTP_PROXY
-    -e HTTPS_PROXY=$HTTPS_PROXY
-    -e no_proxy=$no_proxy
-"
-LOGSETTINGS="--log-driver=syslog --log-opt syslog-facility=local0"
-VOLMAPPING="
-    --privileged -v /dev/bus/usb:/dev/bus/usb
-    -v /tmp/.X11-unix/:/tmp/.X11-unix:Z
-    -v $DATA_DIR/home/liveuser/:/home/liveuser:Z
-"
-
-logger -p local0.info "starting docker image $DOCKER_IMAGE"
-$sudo docker run $runopt --rm \
-    --hostname=$CONTAINERNAME --name=$CONTAINERNAME \
-    $ENVSETTINGS $LOGSETTINGS $VOLMAPPING \
-    $DOCKER_IMAGE
